Replace `NullRelation` with a `#null?` state for `Relation`
In all cases where `ActiveRecord` currently generates a query with
`WHERE 1=0`, this will instead avoid the query and return a `#null?` `Relation`,
ala .none. This covers cases such as `.where('1=0')`, `.where(id: [])`,
and `.limit(0)`.

This line of thought was inspired by the need for #19349, which is
caused `NullRelation` state leaking across `#reset` boundaries. This
alternative is tidier in that it is 100% resetable, and does not
require hacky cache clearing.

This patch would be improved if `Arel::Nodes::In` responded to a method
to indicate when it was a none predicate - see
`WhereClause#none_predicate?`, but I figured I'd get you guys'
opinion on it first.