Print `bundle install` output in `rails new` as soon as it's available
Previously, running `rails new` would not print any of the output from
`bundle install` until all the gems had finished installing. This made
it look like the generator was hanging at the `bundle install` step.

This commit changes how we call the bundle command from using backticks
to using `Open3.popen2e`, allowing us to read each line of the output as
bundler produces it.

This has the added benefit of including output bundler produces on
standard error, which the previous code ignored since backticks only
capture standard out. This is not a big deal right now since bundler
does not currently print errors to standard error, but that may change
in the future (see: bundler/bundler/issues/3353).

The one downside here is that since we're using `Open3.popen2e`, we are
merging standard error and standard out, and outputting them both to
standard out. This is not ideal, but it removes the possibility for a
deadlock that could occur if we used `Open3.popen3` instead.