Check subtype limit before using the default limit
As described here https://github.com/rails/rails/issues/19420. When using the Postgres BigInt[] field type the big int value was not being translated into schema.rb. This caused the field to become just a regular integer field when building off of schema.rb. This fix will addrees this by checking the subtype's limit first and if not found then defaulting to the limit. As in this case the subtype limit is correct.

https://github.com/rails/rails/issues/19420