Fix method signature of `parse_query` to match rack
Recently rack was changed to have a second argument on the `parse_query`
method (in rack/rack#781). Rails relies on this and it's `parse_query`
method was missing the second argument. I've bumped rack in the gemspec
until it can be released so that tests between Rails master and Rack
master pass.