Call destroy not destroy! on dependent children
This is a proposal for a possibly better solution to commit #13042.
Currently, Rails will raise ActiveRecord::RecordNotDestroyed if you call
destroy on a model, but that model has children with dependent: :destroy
set, and those models fail to be destroyed. This causes issues if you
need to perform some action in the parent's before_destroy regardless of
whether the children fail to be destroyed.

This switches back to calling destroy instead of destroy! on the
children. The original issue (#12812) is resolved in this commit by only
raising an ActiveRecord::RecordNotSaved exception when you are replacing
a dependent: :destroy child, and not when calling destroy on the parent.