[WIP] Modify redirect_to to halt further execution.
The purpose of this addition is to prevent possible execution after redirect
exceptions in rails. While the suggested method of handling control flow is to
utililze a before_filter to prevent code in a controller from executing, there
are cases were rails developers will legitimately use `redirect_to` in their
controllers.

With that in mind, there is zero reason for any code, save for ensure blocks,
to execute after a `redirect_to` is issued. Anywhere someone intentionally
executes code following a redirect, it can be restructured to occur prior to
the `redirect_to` exception. If modifing the response body is desired,
`redirect_to` should take a parameter to set a custom response body.

This is reimplementation of my friend and colleague Adam Doupé's prior work:
https://github.com/adamdoupe/rails/commit/78c655015595dc83bd9dccb997d6777f1a51b9cc

Code partially developed on my company's time. Thanks AppFolio!