Fixed lack of HTML safety when content_for is flushed
Previously, when content_for is set to flushed, the value
associated with the key in ActionView::OutputFlow#set was
being replaced entirely. The problem was that the new
value passed to the method may not be an instance of
ActiveSupport::SafeBuffer.

This change forces the value to be appended to a new
instance of ActiveSupport::SafeBuffer.