option to raise on missing template
In 0de4a23 the behavior when there is a missing template was changed to
not raise an error, but instead head :no_content.  This is a breaking
change and some gems rely on this happening.

This adds `raise_on_missing_template` to controllers that, when set to
`true`, enables the old behavior