URL helpers for controller scope routes
Previously, if routes are defined within a
controller scope and an :id parameter is
specified in the route, the mapper fails
to create a default url helper.

Also, according to the docs, the (intended)
old behavior for the url helpers appears to
be that it would be named after the method
specified in the route. For example:

    controller :foo do
      get 'foo/index' => :list
    end

will match '/foo/index' to the 'list_path'.

This seems problematic as default behavior,
as they are not namespaced based on the
controller name. This commit also prepends
the controller name to the url helper, such
that '/foo/index' is matched to
'foo_list_path'.

Documentation was also updated to reflect
this change.

See #20312.