Fix for #20508. Force relation load in empty?/any? methods when having
When a relation has a having clause, empty?/any? methods can not rely
on count(:all) to check if there is any group in the result
set, because count(:all) ignores group/having clauses.

It worked well if group values are present, but it didn't work with
relations including a having clause.

I'm fixing it by executing the query and checking actual results
from DB. Although, we could come up with a different approach if we want
to avoid the full query (e.g. using an exists and putting the relation
sql into a subquery)