Make polymorphic associations joinable by specifying its classes
Polymorphic associations have not been join-able and therefore were
not very memory efficient or fast when alot of data had to be processed
that was associated by a polymorphic.

Since it is also not possible to guess at application boot time which
models might be associated, a good way is to let the developer specify
the associated models that he wants to join eventually.

The join is processed by left outer joining the polymorphic tables.
Inner joins do not work, since they are mutually exclusive on belongs_to
associations and the resultset would therefore be empty.

Example:
  class Treasure < ActiveRecord::Base
    belongs_to :looter, :polymorphic => true,
      :poly_classes => [Parrot, Pirate]
  end

  Ship.joins(:treasures => :looter).where(:parrots => { :name => Mona })