Prevent destructive action on production database
It is possible to run your tests against your production database by accident right now. While infrequently, but as an anecdotal data point, Heroku receives a non-trivial number of requests for a database restore due to this happening. In these cases the loss can be large. 

To prevent against running tests against production we can store the "environment" version that was used when migrating the database in the schema table. Before executing tests we can see if the database is a "production" database and abort. There is a manual escape valve to force this check from happening with environment variable `RUN_AGAINST_PRODUCTION_DATABASE=1`

This is a WIP

TODO

- [] schema load populates environment in the schema table
- [] check for `supports_migrations?`
- [] handle case for when activerecord is included but no database used

To Test

- [] add test for when `rake test` with RAILS_ENV prevents task from running
- [] add test for when `rake test` is called without any environment data in the schema

cc @rafaelfranca @sgrif