Hide overly-aggressive route reloading behind configuration flag
Previously Rails would reload routes on *every* single request in
case a constant used in the routes file(s) had changed. This is
very expensive and introduced a noticeable slowdown in all
parts of the application (application routes, assets, etc.).
This sort of magic isn't worth the cost of such a bad developer
user experience, so until a better solution is found it's
preferable to have it hidden behind a default-off config property.