add test that demonstrates the issue that follows from too much precision on cache_key timestamps.
The default timestamp used for AR is `updated_at` in nanoseconds! (:nsec)
This causes issues on any machine that runs an OS that supports nanoseconds timestamps, i.e. not-OS X.
Where the cache_key of the record persisted in the database (milliseconds precision) is out-of-sync with the cache_key in the ruby VM.

The test is pretty much self-explanatory.