Reveal intent of prefix relative_url_root override
When generating a mounted engine prefix, the value of `#relative_url_root` should be ignored. This value is taken into account in another step of the route generation.

Previous commits set a default `options[:script_name]` to prevent `#relative_url_root` from being called. Instead of a default `prefix_options[:script_name]`, a default `prefix_options[:relative_url_root]` is set to an empty string. This is more intention revealing and leads to the same result of allowing the value of `#relative_url_root` to be conditionally ignored.